backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
datasets==2.14.6
qdrant-client==1.6.0
sentence-transformers==2.2.2
